requirements.txt
streamlit
python-dotenv
requests
openai
graphviz

langchain-core
langchain-openai
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-classic
langchain-chroma
langgraph

chromadb
docx2txt
pypdf
